cleaning and restructured koala
