--- koala/__init__.py
@@ -1,3 +1,0 @@
-import xml
-import excel
-import unzip
