+++ main.py
@@ -9,10 +9,9 @@ import warnings
 from io import BytesIO
 from multiprocessing import Pool
 
-from koala.unzip import read_archive
-from koala.excel.excel import read_named_ranges, read_cells
-from koala.ast.tokenizer import ExcelParser
-from koala.ast.graph import ExcelCompiler, Spreadsheet
+from koala.reader import read_archive, read_named_ranges, read_cells
+from koala.tokenizer import ExcelParser
+from koala import ExcelCompiler, Spreadsheet
